refactor: instantiating into the App class
